+++ tests/test_query.py
@@ -38,7 +38,10 @@ from cray_product_catalog.query import (
     InstalledProductVersion,
     ProductCatalogError
 )
-from tests.mocks import COS_VERSIONS, MOCK_PRODUCT_CATALOG_DATA, SAT_VERSIONS
+from tests.mocks import (
+    COS_VERSIONS, MOCK_PRODUCT_CATALOG_DATA, SAT_VERSIONS,
+    CPE_VERSION, MOCK_INVALID_PRODUCT_DATA
+)
 
 
 class TestGetK8sAPI(unittest.TestCase):
@@ -93,7 +96,7 @@ class TestProductCatalog(unittest.TestCase):
         product_catalog = self.create_and_assert_product_catalog()
         expected_names_and_versions = [
             (name, version) for name in ('sat', 'cos') for version in ('2.0.0', '2.0.1')
-        ] + [('other_product', '2.0.0')]
+        ] + [('cpe', '2.0.0'), ('other_product', '2.0.0')]
         actual_names_and_versions = [
             (product.name, product.version) for product in product_catalog.products
         ]
@@ -112,19 +115,44 @@ class TestProductCatalog(unittest.TestCase):
                                     'No data found in mock-namespace/mock-name ConfigMap.'):
             self.create_and_assert_product_catalog()
 
-    def test_create_product_catalog_invalid_product_schema(self):
-        """Test creating a ProductCatalog when an entry contains valid YAML but does not match schema."""
-        self.mock_k8s_api.read_namespaced_config_map.return_value = Mock(data={
-            'sat': safe_dump({'2.1': {'component_versions': {'docker': 'should be an array'}}})
-        })
+    def check_for_invalid_product_schema(self, prod, ver):
+        """Check for an ProductCatalog entry containing valid YAML but does not match schema.
+        Args:
+            prod (string): Product Name
+            ver (string): Product version
+        """
+        mock_data = {prod: safe_dump(MOCK_INVALID_PRODUCT_DATA.get(prod))}
+        self.mock_k8s_api.read_namespaced_config_map.return_value = Mock(data=mock_data)
+
         with self.assertLogs(level=logging.DEBUG) as logs_cm:
             product_catalog = self.create_and_assert_product_catalog()
 
         self.assertEqual(1, len(logs_cm.records))
-        self.assertEqual('The following products have product catalog data that '
-                         'is not valid against the expected schema: sat-2.1',
-                         logs_cm.records[0].message)
+        message = 'The following products have product catalog data that ' + \
+            'is not valid against the expected schema: ' + prod + '-' + ver
+        self.assertEqual(message, logs_cm.records[0].message)
         self.assertEqual(product_catalog.products, [])
+
+    def test_create_product_catalog_invalid_product_schema_for_docker(self):
+        """
+        Test creating a ProductCatalog when an entry contains valid YAML but does not match schema.
+        As per schema, 'docker' should be an array.
+        """
+        self.check_for_invalid_product_schema('sat', '2.1')
+
+    def test_create_product_catalog_invalid_product_schema_for_s3(self):
+        """
+        Test creating a ProductCatalog when an entry contains valid YAML but does not match schema.
+        As per schema, 's3' should be an array.
+        """
+        self.check_for_invalid_product_schema('cpe', '2.1')
+
+    def test_create_product_catalog_invalid_product_schema_for_manifests(self):
+        """
+        Test creating a ProductCatalog when an entry contains valid YAML but does not match schema.
+        As per schema, 'manifests' should be an array.
+        """
+        self.check_for_invalid_product_schema('cos', '2.1')
 
     def test_get_matching_product(self):
         """Test getting a particular product by name/version."""
@@ -170,25 +198,102 @@ class TestInstalledProductVersion(unittest.TestCase):
         )
 
     def test_no_docker_images(self):
-        """Test a product that has an empty dictionary under the 'docker' key returns an empty dictionary."""
+        """Test a product that has an empty dictionary under the 'docker' key returns an empty list."""
         product_with_no_docker_images = InstalledProductVersion(
             'sat', '0.9.9', {'component_versions': {'docker': {}}}
         )
         self.assertEqual(product_with_no_docker_images.docker_images, [])
 
     def test_no_docker_images_null(self):
-        """Test a product that has None under the 'docker' key returns an empty dictionary."""
+        """Test a product that has None under the 'docker' key returns an empty list."""
         product_with_no_docker_images = InstalledProductVersion(
             'sat', '0.9.9', {'component_versions': {'docker': None}}
         )
         self.assertEqual(product_with_no_docker_images.docker_images, [])
 
     def test_no_docker_images_empty_list(self):
-        """Test a product that has an empty list under the 'docker' key returns an empty dictionary."""
+        """Test a product that has an empty list under the 'docker' key returns an empty list."""
         product_with_no_docker_images = InstalledProductVersion(
             'sat', '0.9.9', {'component_versions': {'docker': []}}
         )
         self.assertEqual(product_with_no_docker_images.docker_images, [])
+
+    def test_helm_charts(self):
+        """Test getting the Helm charts."""
+        expected_helm_charts_versions = [
+            ('cos-config', '0.4.76'),
+            ('cos-sle15sp3-artifacts', '1.3.23'),
+            ('cray-cps', '1.8.15')
+        ]
+        self.assertEqual(
+            expected_helm_charts_versions, self.installed_product_version.helm_charts
+        )
+
+    def test_no_helm_charts(self):
+        """Test a product that has an empty dictionary under the 'helm' key returns an empty list."""
+        product_with_no_helm_charts = InstalledProductVersion(
+            'sat', '0.9.9', {'component_versions': {'helm': {}}}
+        )
+        self.assertEqual(product_with_no_helm_charts.helm_charts, [])
+
+    def test_no_helm_charts_null(self):
+        """Test a product that has None under the 'helm' key returns an empty list."""
+        product_with_no_helm_charts = InstalledProductVersion(
+            'sat', '0.9.9', {'component_versions': {'helm': None}}
+        )
+        self.assertEqual(product_with_no_helm_charts.helm_charts, [])
+
+    def test_no_helm_charts_empty_list(self):
+        """Test a product that has an empty list under the 'helm' key returns an empty list."""
+        product_with_no_helm_charts = InstalledProductVersion(
+            'sat', '0.9.9', {'component_versions': {'helm': []}}
+        )
+        self.assertEqual(product_with_no_helm_charts.helm_charts, [])
+
+    def test_s3_artifacts(self):
+        """Test getting the s3 artifacts."""
+        expected_s3_artifacts = [('boot-images', 'PE/CPE-base.x86_64-2.0.squashfs'),
+                                 ('boot-images', 'PE/CPE-amd.x86_64-2.0.squashfs')]
+        product_with_s3_artifacts = InstalledProductVersion(
+            'cpe', '2.0.0', CPE_VERSION['2.0.0']
+        )
+        self.assertEqual(expected_s3_artifacts, product_with_s3_artifacts.s3_artifacts)
+
+    def test_no_s3_artifacts(self):
+        """Test a product that has an empty dictionary under the 's3' key returns an empty list."""
+        product_with_no_s3_artifacts = InstalledProductVersion(
+            'sat', '0.9.9', {'component_versions': {'s3': {}}}
+        )
+        self.assertEqual(product_with_no_s3_artifacts.s3_artifacts, [])
+
+    def test_no_s3_artifacts_null(self):
+        """Test a product that has None under the 's3' key returns an empty list."""
+        product_with_no_s3_artifacts = InstalledProductVersion(
+            'sat', '0.9.9', {'component_versions': {'s3': None}}
+        )
+        self.assertEqual(product_with_no_s3_artifacts.s3_artifacts, [])
+
+    def test_no_s3_artifacts_empty_list(self):
+        """Test a product that has an empty list under the 's3' key returns an empty list."""
+        product_with_no_s3_artifacts = InstalledProductVersion(
+            'sat', '0.9.9', {'component_versions': {'s3': []}}
+        )
+        self.assertEqual(product_with_no_s3_artifacts.s3_artifacts, [])
+
+    def test_loftsman_manifests(self):
+        """Test getting the Loftsman manifests."""
+        expected_loftsman_manifests = ['config-data/argo/loftsman/cos/2.0.0/manifests/cos-services.yaml']
+        product_with_loftsman_manifests = InstalledProductVersion(
+            'cos', '2.0.0', COS_VERSIONS['2.0.0']
+        )
+        self.assertEqual(expected_loftsman_manifests, product_with_loftsman_manifests.loftsman_manifests)
+
+    def test_no_loftsman_manifests_empty_list(self):
+        """Test a product that has an empty list under the 'manifests' key returns an empty list."""
+        product_with_no_loftsman_manifests = InstalledProductVersion(
+            'sat', '0.9.9', {'component_versions': {'manifests': []}}
+        )
+        self.assertEqual(product_with_no_loftsman_manifests.loftsman_manifests, [])
 
     def test_str(self):
         """Test the string representation of InstalledProductVersion."""